+++ lib/AppInfo/Application.php
@@ -68,6 +68,8 @@ class Application extends App implements IBootstrap {
         $context->registerTaskProcessingProvider(\OCA\EvaAi\TaskProcessing\EvaTextToImageProvider::class);
         $context->registerTaskProcessingProvider(\OCA\EvaAi\TaskProcessing\EvaAudioToTextProvider::class);
         $context->registerTaskProcessingProvider(\OCA\EvaAi\TaskProcessing\EvaAudioSubtitlesProvider::class);
+        $context->registerTaskProcessingProvider(\OCA\EvaAi\TaskProcessing\EvaTextToSpeechProvider::class);
+        $context->registerTaskProcessingProvider(\OCA\EvaAi\TaskProcessing\EvaAudioTranslateProvider::class);
         // Talk-Bot: reagiert auf BotInvokeEvent, wenn Nextcloud Talk installiert ist.
         $context->registerEventListener(\OCA\Talk\Events\BotInvokeEvent::class, \OCA\EvaAi\Listener\TalkBotListener::class);
         // GDPR-Erasure: Kontenloeschung raeumt alle eva_ai-Daten des Users ab (Issue #83).